pyproject.toml
[tool.poetry]
name = "djtutproj"
version = "0.1.0"
description = "trying stuff"
authors = ["Riley <rileymshea@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.3"
psycopg2 = "^2.9.3"
djangorestframework = "^3.13.1"
uritemplate = "^4.1.1"
PyYAML = "^6.0"
django-tailwind = "^3.1.1"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
